Implement Pet Selection Component for personalized nutrition analysis and integrate it into the AIAssistant page. Enhance UserLocationDisplay with improved geolocation handling and error management.
